issue #53 refactor: Fragment 내의 Inner Class 분리 & ReceiptViewModel 호출 Fragment 변경 (CalcMainFragment에서 호출 제외)
