+++ app/src/main/java/com/example/wapapp2/view/calculation/interfaces/OnUpdateSummaryCallback.kt
@@ -1,5 +1,5 @@
 package com.example.wapapp2.view.calculation.interfaces
 
 fun interface OnUpdateSummaryCallback {
-    fun updateSummaryUI()
+    fun updateSummaryUI(summary : Int)
 }
